Pwn overflow1 (#18)
* made a writeupi

* made a writeupi

* fix to readme, rewrite to dump.c

* Update vuln.c to use 8 byte aligned admin variable

* edit per comment request in issue #18

* Update pwn/overflow1/writeup/README.md


* Update README for offsets

---------

+++ pwn/overflow1/src/vuln.c
@@ -41,7 +41,7 @@ int main(){
     ignore_me_timeout();
 
     char name[32];
-    int admin = 0;
+    long int admin = 0;
 
     puts("What is your name?");
     gets(name);
